backend/requirements.txt
fastapi==0.115.12
google-genai==1.18.0
hyperon==0.2.5
python-dotenv==1.1.0
uvicorn==0.34.3
